Container for Chart component
